+++ double/src/cpu/iteration.c
@@ -12,21 +12,21 @@
  *                  padding).
  * @param[out] grid Pointer to the matrix data array to be initialised.
  */
-static void init(const int n, const int m, float* const grid) {
+static void init(const int n, const int m, double* const grid) {
   const int increment = m + 2; // We need to skip the last two columns since
                                // they are copies of the first and second column
 
   // Loop over rows
   for (int i = 0; i < n; i++) {
-    float col0 =
-        0.98f * (float) ((i + 1) * (i + 1)) /
-        (float) (n * n); // Set it as a variable to make initialisation easier
+    double col0 =
+        0.98 * (double) ((i + 1) * (i + 1)) /
+        (double) (n * n); // Set it as a variable to make initialisation easier
     grid[i * increment + 0] = col0; // We have a '+ 0' here for consistency
 
     // Interior points
     for (int j = 1; j < m; j++) {
       grid[i * increment + j] =
-          col0 * ((float) (m - j) * (m - j) / (float) (m * m));
+          col0 * ((double) (m - j) * (m - j) / (double) (m * m));
     }
 
     // Extra columns
@@ -49,31 +49,31 @@ static void init(const int n, const int m, float* const grid) {
  * @param[out] dst Pointer to the destination row array (m + 2 elements).
  * @param[in]  src Pointer to the source row array (m + 2 elements).
  */
-static void iteration(const int m, float* const restrict dst,
-                      const float* const restrict src) {
+static void iteration(const int m, double* const restrict dst,
+                      const double* const restrict src) {
 
   // Here, we deal with the column at j = 1, since it also wraps around
   {
     int j = 1;
 
     // Left neighbours
-    float old_l2 = src[m - 1];
-    float old_l1 = src[0];
+    double old_l2 = src[m - 1];
+    double old_l1 = src[0];
 
     // Right neighbours
-    float old_r1 = src[j + 1];
-    float old_r2 = src[j + 2];
+    double old_r1 = src[j + 1];
+    double old_r2 = src[j + 2];
 
-    dst[j] = ((1.60f * old_l2) + (1.55f * old_l1) + src[j] + (0.60f * old_r1) +
-              (0.25f * old_r2)) /
-             5.0f;
+    dst[j] = ((1.60 * old_l2) + (1.55 * old_l1) + src[j] + (0.60 * old_r1) +
+              (0.25 * old_r2)) /
+             5.0;
   }
 
   // Other updates
   for (int j = 2; j < m; j++) {
-    dst[j] = ((1.60f * src[j - 2]) + (1.55f * src[j - 1]) + src[j] +
-              (0.60f * src[j + 1]) + (0.25f * src[j + 2])) /
-             5.0f;
+    dst[j] = ((1.60 * src[j - 2]) + (1.55 * src[j - 1]) + src[j] +
+              (0.60 * src[j + 1]) + (0.25 * src[j + 2])) /
+             5.0;
   }
 
   // Refresh our extra columns for further iterations
@@ -94,8 +94,8 @@ static void iteration(const int m, float* const restrict dst,
  * @param[in,out] dst   Pointer to one buffer for the row data.
  * @param[in,out] src   Pointer to the other buffer for the row data.
  */
-static void iterations(const int iters, const int m, float* restrict dst,
-                       float* restrict src) {
+static void iterations(const int iters, const int m, double* restrict dst,
+                       double* restrict src) {
   if (iters % 2 != 0) {
     iteration(m, dst, src);
   }
@@ -122,7 +122,7 @@ static void iterations(const int iters, const int m, float* restrict dst,
  * @param[in,out] src   Pointer to the other buffer for the matrix data.
  */
 void heat_propagation(const int iters, const int n, const int m,
-                      float* restrict dst, float* restrict src) {
+                      double* restrict dst, double* restrict src) {
   const int increment = m + 2;
 
   // Set initial conditions
@@ -130,8 +130,8 @@ void heat_propagation(const int iters, const int n, const int m,
   init(n, m, src);
 
   for (int i = 0; i < n; i++) {
-    float* const row_dst = dst + i * increment;
-    float* const row_src = src + i * increment;
+    double* const row_dst = dst + i * increment;
+    double* const row_src = src + i * increment;
     iterations(iters, m, row_dst, row_src); // Perform the iterations
   }
   return;
